First pass at multi-build support [ci skip]
